Fixes #3875. reverts [14292]
The official faad.h has unsigned long int in it's function declarations
so original code is correct.
Some linux Distros (Gentoo, Debian, ...) patch the public API to use
stdint.h types. There is some code to handle both gracefully, which might
be not in all cases correct.


git-svn-id: http://svn.mythtv.org/svn/trunk@14302 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mythmusic/mythmusic/aacdecoder.h
@@ -15,6 +15,7 @@
 
 #include <mp4ff.h>
 
+
 class Metadata;
 
 class aacDecoder : public Decoder
@@ -50,7 +51,7 @@ class aacDecoder : public Decoder
     bool done, finish;
     long len, bitrate;
     uchar channels;
-    uint32_t sample_rate;
+    unsigned long sample_rate;
     unsigned long output_size;
     double totalTime, seekTime;
